Rewrite demo snippets.
