--- src/Neptuo.Productivity.SnippetManager.UI/Views/DesignData/DemoSnippetProvider.cs
@@ -1,51 +1,0 @@
-﻿using Neptuo.Productivity.SnippetManager.Models;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Neptuo.Productivity.SnippetManager.Views.DesignData;
-
-internal class DemoSnippetProvider : ISnippetProvider
-{
-    public Task InitializeAsync(SnippetProviderContext context) 
-        => Task.CompletedTask;
-
-    public Task UpdateAsync(SnippetProviderContext context)
-    {
-        void Add(string title, string text, string? description = null)
-            => context.Snippets.Add(new SnippetModel(title, text, description));
-
-        Add(
-            "C# class",
-            """
-            using System;
-            using System.Collections.Generic;
-            using System.IO;
-            using System.Linq;
-            using System.Text;
-            using System.Threading.Tasks;
-
-            namespace Namespace
-            {
-                public class Class
-                {
-                }
-            }
-            """,
-            "public class Class..."
-        );
-        Add("Google", "https://google.com");
-        Add("Maps", "https://maps.google.com");
-        Add("GitHub - dotnet - runtime", "https://github.com/dotnet/runtime");
-        Add("GitHub - Maraf - Money", "https://github.com/maraf/money");
-        Add("GitHub - Neptuo - Recollections", "https://github.com/neptuo/Recollections");
-        Add("Money", "https://app.money.neptuo.com");
-        Add("Recollections", "https://recollections.app");
-        Add("Signature", $"S pozdravem{Environment.NewLine}Marek Fišera");
-
-        return Task.CompletedTask;
-    }
-}
