+++ src/Neptuo.Productivity.SnippetManager.UI/Views/DesignData/ViewModelLocator.cs
@@ -1,4 +1,5 @@
-﻿using Neptuo.Productivity.SnippetManager.ViewModels;
+﻿using Neptuo.Productivity.SnippetManager.Models;
+using Neptuo.Productivity.SnippetManager.ViewModels;
 using Neptuo.Productivity.SnippetManager.ViewModels.Commands;
 using System;
 using System.Collections.Generic;
@@ -31,6 +32,37 @@ internal class ViewModelLocator
         }
     }
 
-    private static async void LoadSnippets() 
-        => MainViewModel.Snippets.AddRange(await new DemoSnippetProvider().GetAsync());
+    private static void LoadSnippets()
+    {
+        void Add(string title, string text, string? description = null)
+            => MainViewModel.Snippets.Add(new SnippetModel(title, text, description));
+
+        Add(
+            "C# class",
+            """
+            using System;
+            using System.Collections.Generic;
+            using System.IO;
+            using System.Linq;
+            using System.Text;
+            using System.Threading.Tasks;
+
+            namespace Namespace
+            {
+                public class Class
+                {
+                }
+            }
+            """,
+            "public class Class..."
+        );
+        Add("Google", "https://google.com");
+        Add("Maps", "https://maps.google.com");
+        Add("GitHub - dotnet - runtime", "https://github.com/dotnet/runtime");
+        Add("GitHub - Maraf - Money", "https://github.com/maraf/money");
+        Add("GitHub - Neptuo - Recollections", "https://github.com/neptuo/Recollections");
+        Add("Money", "https://app.money.neptuo.com");
+        Add("Recollections", "https://recollections.app");
+        Add("Signature", $"S pozdravem{Environment.NewLine}Marek Fišera");
+    }
 }
